+++ packages/eslint-config-vtex/rules/style.js
@@ -1,9 +1,10 @@
 module.exports = {
   rules: {
     // Allow brace-less single-line if, else if, else, for, while, or do, while still enforcing the use of curly braces for other instances.
-    // ! Overrides definition on 'eslint-config-prettier'
     // https://eslint.org/docs/rules/curly
-    curly: ['error', 'multi-line', 'consistent'],
+    // Disabled because conflicts with prettier
+    // https://github.com/prettier/eslint-config-prettier#curly
+    curly: 'off',
 
     // Require camel case names
     // https://eslint.org/docs/rules/camelcase